requirements.txt
anyio==3.6.2
beautifulsoup4==4.11.2
bs4==0.0.1
certifi==2022.12.7
charset-normalizer==3.0.1
click==8.1.3
colorama==0.4.6
ecdsa==0.18.0
fastapi==0.89.1
h11==0.14.0
httpcore==0.16.3
httpx==0.23.3
idna==3.4
pyasn1==0.4.8
pydantic==1.10.4
python-jose==3.3.0
requests==2.28.2
rfc3986==1.5.0
rsa==4.9
six==1.16.0
sniffio==1.3.0
soupsieve==2.3.2.post1
starlette==0.22.0
typing-extensions==4.4.0
urllib3==1.26.14
uvicorn==0.20.0
